split the course_db into funcs
