+++ internal/repository/postgres/course_db_navigation.go
@@ -5,178 +5,10 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-
 	coursemodels "skillForce/internal/models/course"
 	"skillForce/internal/models/dto"
 	"skillForce/pkg/logs"
 )
-
-func (d *Database) GetBucketCourses(ctx context.Context) ([]*coursemodels.Course, error) {
-	//TODO: можно заморочиться и сделать самописную пагинацию через LIMIT OFFSET
-	var bucketCourses []*coursemodels.Course
-	rows, err := d.conn.Query("SELECT id, creator_user_id, title, description, avatar_src, price, time_to_pass FROM course LIMIT 16")
-	if err != nil {
-		logs.PrintLog(ctx, "GetBucketCourses", fmt.Sprintf("%+v", err))
-		return nil, err
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var course coursemodels.Course
-		if err := rows.Scan(&course.Id, &course.CreatorId, &course.Title, &course.Description, &course.ScrImage, &course.Price, &course.TimeToPass); err != nil {
-			logs.PrintLog(ctx, "GetBucketCourses", fmt.Sprintf("%+v", err))
-			return nil, err
-		}
-		logs.PrintLog(ctx, "GetBucketCourses", fmt.Sprintf("get course %+v from db", course))
-		bucketCourses = append(bucketCourses, &course)
-	}
-
-	logs.PrintLog(ctx, "GetBucketCourses", "get bucket ourses from db")
-
-	return bucketCourses, nil
-}
-
-func (d *Database) GetCoursesPurchases(ctx context.Context, bucketCoursesWithoutPurchases []*coursemodels.Course) (map[int]int, error) {
-	coursesPurchases := make(map[int]int, len(bucketCoursesWithoutPurchases))
-
-	for _, course := range bucketCoursesWithoutPurchases {
-		var purchases int
-		err := d.conn.QueryRow("SELECT COUNT(*) FROM SIGNUPS WHERE course_id = $1", course.Id).Scan(&purchases)
-		if err != nil {
-			logs.PrintLog(ctx, "GetBucketCoursesPurchases", fmt.Sprintf("%+v", err))
-			return nil, err
-		}
-
-		coursesPurchases[course.Id] = purchases
-	}
-	return coursesPurchases, nil
-}
-
-func (d *Database) GetCoursesRaitings(ctx context.Context, bucketCoursesWithoutRating []*coursemodels.Course) (map[int]float32, error) {
-	coursesRatings := make(map[int]float32, len(bucketCoursesWithoutRating))
-
-	for _, course := range bucketCoursesWithoutRating {
-		rows, err := d.conn.Query("SELECT rating FROM course_metrik WHERE course_id = $1", course.Id)
-		if err != nil {
-			logs.PrintLog(ctx, "GetCoursesRaitings", fmt.Sprintf("%+v", err))
-			return nil, err
-		}
-		defer rows.Close()
-
-		var sumMetrics float32
-		var countMetrics float32
-
-		for rows.Next() {
-			var metric float32
-			if err := rows.Scan(&metric); err != nil {
-				logs.PrintLog(ctx, "GetCoursesRaitings", fmt.Sprintf("%+v", err))
-				return nil, err
-			}
-			sumMetrics += metric
-			countMetrics++
-		}
-
-		if countMetrics == 0 {
-			continue
-		}
-
-		coursesRatings[course.Id] = sumMetrics / countMetrics
-	}
-
-	logs.PrintLog(ctx, "GetCoursesRaitings", "get courses ratings from db")
-	return coursesRatings, nil
-}
-
-func (d *Database) GetCoursesTags(ctx context.Context, bucketCoursesWithoutTags []*coursemodels.Course) (map[int][]string, error) {
-	coursesTags := make(map[int][]string, len(bucketCoursesWithoutTags))
-
-	for _, course := range bucketCoursesWithoutTags {
-		rows, err := d.conn.Query(`
-			SELECT vt.Title
-			FROM TAGS t
-			JOIN VALID_TAGS vt ON t.Tag_ID = vt.ID
-			WHERE t.Course_ID = $1
-		`, course.Id)
-
-		if err != nil {
-			logs.PrintLog(ctx, "GetCoursesTags", fmt.Sprintf("%+v", err))
-			return nil, err
-		}
-		defer rows.Close()
-
-		var tags []string
-
-		for rows.Next() {
-			var tag string
-			if err := rows.Scan(&tag); err != nil {
-				logs.PrintLog(ctx, "GetCoursesTags", fmt.Sprintf("%+v", err))
-				return nil, err
-			}
-			tags = append(tags, tag)
-		}
-
-		if len(tags) == 0 {
-			continue
-		}
-
-		coursesTags[course.Id] = tags
-
-	}
-	logs.PrintLog(ctx, "GetCoursesTags", "get courses tags from db")
-	return coursesTags, nil
-}
-
-func (d *Database) GetCourseById(ctx context.Context, courseId int) (*coursemodels.Course, error) {
-	var course coursemodels.Course
-	err := d.conn.QueryRow("SELECT id, creator_user_id, title, description, avatar_src, price, time_to_pass FROM course WHERE id = $1", courseId).Scan(
-		&course.Id, &course.CreatorId, &course.Title, &course.Description, &course.ScrImage, &course.Price, &course.TimeToPass)
-	if err != nil {
-		logs.PrintLog(ctx, "GetCourseById", fmt.Sprintf("%+v", err))
-		return nil, err
-	}
-	logs.PrintLog(ctx, "GetCourseById", fmt.Sprintf("get course %+v from db", course))
-	return &course, nil
-}
-
-func (d *Database) MarkLessonCompleted(ctx context.Context, userId int, courseId int, lessonId int) error {
-	exists := false
-	err := d.conn.QueryRow("SELECT EXISTS (SELECT 1 FROM LESSON_CHECKPOINT WHERE user_id = $1 AND lesson_id = $2 AND course_id = $3)",
-		userId, lessonId, courseId).Scan(&exists)
-	if err != nil {
-		logs.PrintLog(ctx, "MarkLessonComplete", fmt.Sprintf("%+v", err))
-	}
-
-	if exists {
-		logs.PrintLog(ctx, "MarkLessonComplete", fmt.Sprintf("lesson id:%+v is already learned by the user id:%+v", lessonId, userId))
-		return nil
-	}
-
-	_, err = d.conn.Exec(
-		"INSERT INTO LESSON_CHECKPOINT (user_id, lesson_id, course_id) VALUES ($1, $2, $3)",
-		userId, lessonId, courseId)
-
-	if err != nil {
-		logs.PrintLog(ctx, "MarkLessonComplete", fmt.Sprintf("%+v", err))
-		return err
-	}
-
-	logs.PrintLog(ctx, "MarkLessonComplete", fmt.Sprintf("mark that lesson id:%+v is learned by the user id:%+v", lessonId, userId))
-	return nil
-}
-
-func (d *Database) MarkLessonAsNotCompleted(ctx context.Context, userId int, lessonId int) error {
-	_, err := d.conn.Exec(
-		"DELETE FROM LESSON_CHECKPOINT WHERE user_id = $1 AND lesson_id = $2",
-		userId, lessonId)
-
-	if err != nil {
-		logs.PrintLog(ctx, "markLessonAsNotComplete", fmt.Sprintf("%+v", err))
-		return err
-	}
-
-	logs.PrintLog(ctx, "markLessonAsNotComplete", fmt.Sprintf("mark that lesson id:%+v is not learned by the user id:%+v", lessonId, userId))
-	return nil
-}
 
 func (d *Database) getLessonHeaderNewCourse(ctx context.Context, userId int, courseId int) (*dto.LessonDtoHeader, int, string, bool, error) {
 	var part coursemodels.CoursePart
@@ -498,78 +330,6 @@ func (d *Database) GetLessonHeaderByLessonId(ctx context.Context, userId int, cu
 	return lessonHeader, nil
 }
 
-func (d *Database) GetLessonBlocks(ctx context.Context, currentLessonId int) ([]string, error) {
-	var blocks []string
-	rows, err := d.conn.Query(`
-			SELECT tlb.value
-			FROM TEXT_LESSON_BLOCK tlb
-			JOIN TEXT_LESSON tl ON tlb.Text_Lesson_ID = tl.ID
-			WHERE tl.Lesson_ID = $1
-			ORDER BY tlb.Text_Lesson_Block_Order ASC
-		`, currentLessonId)
-	if err != nil {
-		logs.PrintLog(ctx, "GetLessonBlocks", fmt.Sprintf("%+v", err))
-		return nil, err
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var block string
-		if err := rows.Scan(&block); err != nil {
-			logs.PrintLog(ctx, "GetLessonBlocks", fmt.Sprintf("%+v", err))
-			return nil, err
-		}
-		blocks = append(blocks, block)
-	}
-	return blocks, nil
-}
-
-func (d *Database) GetLessonVideo(ctx context.Context, currentLessonId int) ([]string, error) {
-	var videoSrc string
-	err := d.conn.QueryRow(`
-			SELECT video_src
-			FROM VIDEO_LESSON
-			WHERE lesson_ID = $1
-		`, currentLessonId).Scan(&videoSrc)
-	if err != nil {
-		logs.PrintLog(ctx, "GetLessonVideos", fmt.Sprintf("%+v", err))
-		return nil, err
-	}
-	return []string{videoSrc}, nil
-}
-
-func (d *Database) GetLessonById(ctx context.Context, lessonId int) (*coursemodels.LessonPoint, error) {
-	var lesson coursemodels.LessonPoint
-	err := d.conn.QueryRow(`
-			SELECT id, title, type
-			FROM LESSON
-			WHERE id = $1
-		`, lessonId).Scan(&lesson.LessonId, &lesson.Title, &lesson.Type)
-	if err != nil {
-		logs.PrintLog(ctx, "GetLessonById", fmt.Sprintf("%+v", err))
-		return nil, err
-	}
-	return &lesson, nil
-}
-
-func (d *Database) GetBucketByLessonId(ctx context.Context, currentLessonId int) (*coursemodels.LessonBucket, error) {
-	var bucketId int
-	err := d.conn.QueryRow(`
-			SELECT lesson_bucket_id
-			FROM LESSON
-			WHERE id = $1
-		`, currentLessonId).Scan(&bucketId)
-	if err != nil {
-		logs.PrintLog(ctx, "GetBucketByLessonId", fmt.Sprintf("%+v", err))
-		return nil, err
-	}
-
-	bucket := &coursemodels.LessonBucket{
-		Id: bucketId,
-	}
-	return bucket, nil
-}
-
 func (d *Database) GetLessonFooters(ctx context.Context, currentLessonId int) ([]int, error) {
 	footers := []int{-1, -1, -1}
 
@@ -661,159 +421,4 @@ func (d *Database) GetLessonFooters(ctx context.Context, currentLessonId int) ([
 	logs.PrintLog(ctx, "GetLessonFooters", fmt.Sprintf("footers: %+v", footers))
 
 	return footers, nil
-}
-
-func (d *Database) GetCourseParts(ctx context.Context, courseId int) ([]*coursemodels.CoursePart, error) {
-	var courseParts []*coursemodels.CoursePart
-	rows, err := d.conn.Query(`
-			SELECT id, title
-			FROM PART
-			WHERE course_id = $1
-			ORDER BY part_order ASC
-		`, courseId)
-	if err != nil {
-		logs.PrintLog(ctx, "GetCourseParts", fmt.Sprintf("%+v", err))
-		return nil, err
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var coursePart coursemodels.CoursePart
-		if err := rows.Scan(&coursePart.Id, &coursePart.Title); err != nil {
-			logs.PrintLog(ctx, "GetCourseParts", fmt.Sprintf("%+v", err))
-			return nil, err
-		}
-		logs.PrintLog(ctx, "GetCourseParts", fmt.Sprintf("get course part %+v", coursePart))
-		courseParts = append(courseParts, &coursePart)
-	}
-	return courseParts, nil
-}
-
-func (d *Database) GetPartBuckets(ctx context.Context, partId int) ([]*coursemodels.LessonBucket, error) {
-	var buckets []*coursemodels.LessonBucket
-	rows, err := d.conn.Query(`
-			SELECT id, title
-			FROM LESSON_BUCKET
-			WHERE part_id = $1
-			ORDER BY lesson_bucket_order ASC
-		`, partId)
-	if err != nil {
-		logs.PrintLog(ctx, "GetPartBuckets", fmt.Sprintf("%+v", err))
-		return nil, err
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var bucket coursemodels.LessonBucket
-		if err := rows.Scan(&bucket.Id, &bucket.Title); err != nil {
-			logs.PrintLog(ctx, "GetPartBuckets", fmt.Sprintf("%+v", err))
-			return nil, err
-		}
-		logs.PrintLog(ctx, "GetPartBuckets", fmt.Sprintf("get bucket %+v", bucket))
-		buckets = append(buckets, &bucket)
-	}
-	return buckets, nil
-}
-
-func (d *Database) GetBucketLessons(ctx context.Context, userId int, courseId int, bucketId int) ([]*coursemodels.LessonPoint, error) {
-	completedLessons := make(map[int]bool)
-	rows1, err := d.conn.Query(`
-			SELECT lesson_id
-			FROM LESSON_CHECKPOINT
-			WHERE user_id = $1 AND course_id = $2
-		`, userId, courseId)
-	if err != nil {
-		logs.PrintLog(ctx, "GetBucketLessons", fmt.Sprintf("%+v", err))
-		return nil, err
-	}
-	defer rows1.Close()
-
-	for rows1.Next() {
-		var completedLessonId int
-		if err := rows1.Scan(&completedLessonId); err != nil {
-			logs.PrintLog(ctx, "GetBucketLessons", fmt.Sprintf("%+v", err))
-			return nil, err
-		}
-		completedLessons[completedLessonId] = true
-	}
-
-	var lessons []*coursemodels.LessonPoint
-	rows2, err := d.conn.Query(`
-			SELECT id, title, type
-			FROM LESSON
-			WHERE lesson_bucket_id = $1
-			ORDER BY lesson_order ASC
-		`, bucketId)
-	if err != nil {
-		logs.PrintLog(ctx, "GetBucketLessons", fmt.Sprintf("%+v", err))
-		return nil, err
-	}
-	defer rows2.Close()
-
-	for rows2.Next() {
-		var lesson coursemodels.LessonPoint
-		if err := rows2.Scan(&lesson.LessonId, &lesson.Title, &lesson.Type); err != nil {
-			logs.PrintLog(ctx, "GetBucketLessons", fmt.Sprintf("%+v", err))
-			return nil, err
-		}
-
-		if _, ok := completedLessons[lesson.LessonId]; ok {
-			lesson.IsDone = true
-		}
-
-		logs.PrintLog(ctx, "GetBucketLessons", fmt.Sprintf("get lesson %+v", lesson))
-		lessons = append(lessons, &lesson)
-	}
-	return lessons, nil
-}
-
-func (d *Database) AddUserToCourse(ctx context.Context, userId int, courseId int) error {
-	exists := false
-	err := d.conn.QueryRow("SELECT EXISTS (SELECT 1 FROM SIGNUPS WHERE user_id = $1 AND course_id = $2)",
-		userId, courseId).Scan(&exists)
-	if err != nil {
-		logs.PrintLog(ctx, "AddUserToCourse", fmt.Sprintf("%+v", err))
-	}
-
-	if exists {
-		logs.PrintLog(ctx, "AddUserToCourse", fmt.Sprintf("user with id %+v is already in course with id %+v", userId, courseId))
-		return nil
-	}
-
-	_, err = d.conn.Exec(
-		"INSERT INTO SIGNUPS (user_id, course_id) VALUES ($1, $2)",
-		userId, courseId)
-	if err != nil {
-		logs.PrintLog(ctx, "AddUserToCourse", fmt.Sprintf("%+v", err))
-		return err
-	}
-	logs.PrintLog(ctx, "AddUserToCourse", fmt.Sprintf("add user with id %+v to course with id %+v", userId, courseId))
-	return nil
-}
-
-func (d *Database) GetVideoUrl(ctx context.Context, lessonId int) (string, error) {
-	var videoUrl string
-	err := d.conn.QueryRow("SELECT video_src FROM video_lesson WHERE lesson_id = $1", lessonId).Scan(&videoUrl)
-	if err != nil {
-		logs.PrintLog(ctx, "GetVideo", fmt.Sprintf("%+v", err))
-		return "", err
-	}
-	return videoUrl, nil
-}
-
-func (d *Database) IsUserPurchasedCourse(ctx context.Context, userId int, courseId int) (bool, error) {
-	var exists bool
-	err := d.conn.QueryRow("SELECT EXISTS (SELECT 1 FROM SIGNUPS WHERE user_id = $1 AND course_id = $2)",
-		userId, courseId).Scan(&exists)
-
-	if errors.Is(err, sql.ErrNoRows) {
-		logs.PrintLog(ctx, "IsUserPurchasedCourse", fmt.Sprintf("%+v", err))
-		return false, nil
-	}
-
-	if err != nil {
-		logs.PrintLog(ctx, "IsUserPurchasedCourse", fmt.Sprintf("%+v", err))
-		return false, err
-	}
-	return exists, nil
 }
